Revert "merging with cvs autotools branch"
This reverts commit bc9f97485e3925e3b267b77776dc114b93b6167c.
and all modifs from seb since bc9f974

Conflicts:

	mc/wscript

+++ src/starmatch.cc
@@ -144,8 +144,8 @@ void StarMatchList::RefineTransfo(const double &NSigmas)
 /* not very robust : assumes that we went through Refine just before... */
 double StarMatchList::Residual() const
 {
-  double deno = (2.*size() - transfo->Npar());
-  return (deno>0.) ? sqrt(dist2/deno) : -1; // is -1 a good idea?
+  int deno = (2.*size() - transfo->Npar());
+  return (deno>0) ? sqrt(dist2/deno) : -1; // is -1 a good idea?
 }
